Refactor StreamingPage to include code from LiveMap and pass LiveStream to ViewerPage.
